requirements.txt
RPi.GPIO==0.7.1
smbus==1.1.post2
openai==1.30.1
pillow==10.3.0
python-dotenv==1.0.1
numpy==2.0.0
pygame==2.6.0